Add owner and access control

+++ gifting/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.core.paginator import Paginator
 from django.contrib import messages
 from django.utils.translation import gettext_lazy as _
+from django.contrib.auth.decorators import login_required
 
 from .models import Ad
 from .filters import AdFilter
@@ -26,14 +27,16 @@ def detail(request, pk):
     return render(request, "gifting/ad_detail.html", {"ad": ad})
 
 
+@login_required
 def create(request):
     """Create a new Ad"""
     form = AdForm()
     if request.method == "POST":
         form = AdForm(request.POST)
         if form.is_valid():
+            ad = form.save(commit=False)
+            ad.owner = request.user
             form.save()
-            ad = form.instance
             messages.success(
                 request, _("%(ad_type)s created") % {"ad_type": ad.get_type_display()}
             )
@@ -44,6 +47,9 @@ def create(request):
 def edit(request, pk):
     """Edit an existing Ad"""
     ad = get_object_or_404(Ad, pk=pk)
+    if not ad.owner == request.user:
+        return redirect("/users/login?next=%s" % request.path)
+
     form = AdForm(instance=ad)
     if request.method == "POST":
         form = AdForm(request.POST, instance=ad)
@@ -57,6 +63,9 @@ def edit(request, pk):
 
 def delete(request, pk):
     ad = get_object_or_404(Ad, pk=pk)
+    if not ad.owner == request.user:
+        return redirect("/users/login?next=%s" % request.path)
+
     if request.method == "POST":
         ad.delete()
         messages.success(
